Feat: make allmost all settings configureable in the ui
- all settings are configureable in mympd.conf or by environment variables
- most settings are overwriteable trough the ui

+++ src/list.h
@@ -28,7 +28,7 @@ struct node {
 };
 
 struct list {
-    unsigned length;
+    long length;
     struct node *list;
 };
